Moved AudioPlayer into a separate file (component)
